+++ StripePaymentSheet/StripePaymentSheet/Source/PaymentSheet/ViewControllers/PaymentSheetVerticalViewController.swift
@@ -216,9 +216,8 @@ class PaymentSheetVerticalViewController: UIViewController, FlowControllerViewCo
                     }
                 case .new(let confirmParams):
                     return confirmParams
-                case .link:
-                    // Handled elsewhere
-                    return nil
+                case .link(let confirmOption):
+                    return confirmOption.signupConfirmParams
                 case .applePay, .external:
                     return nil
                 }
@@ -847,6 +846,8 @@ extension PaymentSheetVerticalViewController: VerticalPaymentMethodListViewContr
                 return confirmParams
             } else if case let .saved(_, confirmParams) = previousPaymentOption {
                 return confirmParams
+            } else if case let .link(confirmOption) = previousPaymentOption {
+                return confirmOption.signupConfirmParams
             } else {
                 return nil
             }